Use a StubbedDataCollection object and share between MockService and XCUI testing
