+++ Sources/Sham/Extensions/MockService+LaunchEnvironment.swift
@@ -4,18 +4,15 @@ import Foundation
 
 public extension MockService {
     /// Sets up the mock service object using launch environment mocked data information.
-    func setupWithLaunchEnvironment() {
-        // Find and decode the launch environment mocked data.
+    static func setupWithLaunchEnvironment() {
+        // Find and decode the launch environment stubbed data collection.
         guard
-            let mockDataObjectString = ProcessInfo.processInfo.environment[LaunchEnvironmentMockData.launchEnvironmentKey],
+            let mockDataObjectString = ProcessInfo.processInfo.environment[StubbedDataCollection.launchEnvironmentKey],
             let mockDataObjectData = mockDataObjectString.data(using: .utf8),
-            let mockDataObject = try? JSONDecoder().decode(LaunchEnvironmentMockData.self, from: mockDataObjectData) else {
+            let stubbedDataCollection = try? JSONDecoder().decode(StubbedDataCollection.self, from: mockDataObjectData) else {
             return
         }
         
-        // Populate the mock service with the stubbed data.
-        mockDataObject.stubbedData.forEach { key, value in
-            self.stub(key, with: value)
-        }
+        MockService.shared = MockService(withCollection: stubbedDataCollection)
     }
 }
