+++ Sources/Sham/MockService.swift
@@ -8,7 +8,7 @@ public class MockService {
     // MARK: - Shared Instance
     
     /// The shared instance of the service.
-    public static let shared = MockService()
+    public static var shared = MockService()
     
     // MARK: - Properties
     
@@ -29,11 +29,11 @@ public class MockService {
     }()
     
     /// A dictionary of stubbed responses keyed by stubbed requests.
-    private var stubbedData = [StubRequest: StubResponse]()
+    private(set) var stubbedDataCollection = StubbedDataCollection()
     
     /// Whether or not there are any stubbed response.
     public var hasStubs: Bool {
-        return !self.stubbedData.isEmpty
+        return !self.stubbedDataCollection.stubbedData.isEmpty
     }
     
     // MARK: - Methods
@@ -44,6 +44,12 @@ public class MockService {
     /// using any instance of MockService other than the shared instance would cause problems.
     /// Until we think of a way to allow this, we're restricting this class to just a shared instance for now.
     private init() {}
+    
+    /// Initalizes a MockService with a given stubbed data collection.
+    /// - Parameter collection: The stubbed data collection to use.
+    init(withCollection collection: StubbedDataCollection) {
+        self.stubbedDataCollection = collection
+    }
     
     // MARK: Registration
     
@@ -67,33 +73,7 @@ public class MockService {
     /// Returns nil if there is no matching stubbed request found.
     /// - Parameter request: The request to match against stubbed requests.
     public func getResponse(for request: StubRequest) -> StubResponse? {
-        self.log("Attempting to stub request: \(request.description)")
-        
-        // Check for a match with the exact URL
-        var response = self.stubbedData[request]
-        
-        // If response had no exact match, we need to find the closest match
-        if response == nil {
-            // Filter out stubs that can not mock this request
-            let availableStubs = self.stubbedData.filter {
-                return $0.key.canMockData(for: request)
-            }
-            
-            // Find the highest priority matching score to get the best response
-            response = availableStubs.max {
-                $0.key.priorityScore(for: request) < $1.key.priorityScore(for: request)
-            }?.value
-        }
-        
-        // Log response debugging information
-        if let stubResponse = response {
-            self.log("Found stub response: \(stubResponse)")
-        } else {
-            self.log("No stub response found.")
-        }
-        
-        // Return the response
-        return response
+        return self.stubbedDataCollection.getResponse(for: request)
     }
     
     /// Determines whether or not a matching request has been stubbed.
@@ -113,18 +93,7 @@ public class MockService {
     /// - Parameter request: The request to stub.
     /// - Parameter response: The response to return upon receiving the given request.
     public func stub(_ request: StubRequest, with response: StubResponse) {
-        // Ensure that the URL is valid for stubbing
-        guard request.isValidForStubbing else {
-            // TODO: Throw error
-            return
-        }
-        
-        // If the request exists, print a log to verify
-        if self.stubbedData.keys.contains(request) {
-            self.log("Stubbed data already exists for request '\(request)'. Updating with new data.")
-        }
-        
-        self.stubbedData[request] = response
+        self.stubbedDataCollection.stub(request, with: response)
     }
     
     // MARK: URLRequest Convenience
@@ -163,7 +132,7 @@ public class MockService {
     
     /// Clears all stubbed responses from the stub response collection.
     public func clearData() {
-        self.stubbedData = [:]
+        self.stubbedDataCollection.clearData()
     }
     
     private func log(_ message: String) {
